services/api/pyproject.toml
[project]
name = "boldclub-api"
version = "0.1.0"
description = "FastAPI backend for BoldClub"
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.29,<1.0",
  "aiosqlite>=0.20,<1.0",
  "celery>=5.4,<6.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "openai>=1.68,<2.0",
  "PyJWT>=2.9,<3.0",
  "pytest>=8.2,<9.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "redis>=5.0,<6.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.30,<1.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
